Move models into their own files, remove DB.go
