+++ go-battle/.air.toml
@@ -5,7 +5,7 @@ tmp_dir = "tmp"
 [build]
   args_bin = []
   bin = "./bin/api"
-  cmd = "go build -o bin/api api.go Match.go Player.go Game.go Client.go DB.go"
+  cmd = "go build -o bin/api ."
   delay = 1000
   exclude_dir = ["assets", "tmp", "vendor", "testdata"]
   exclude_file = []
