+++ go-battle/GameClient.go
@@ -1,0 +1,17 @@
+package main
+
+import "gorm.io/gorm"
+
+// GameClient represents a client in a game
+type GameClient struct {
+	gorm.Model
+
+	Name         string `json:"name"`
+	Index        int    `json:"index"`
+	Spectating   bool   `json:"spectating"`
+	Won          bool   `json:"won"`
+	Lost         bool   `json:"lost"`
+	Reason       string `json:"reason"`
+	Disconnected bool   `json:"disconnected"`
+	TimedOut     bool   `json:"timedOut"`
+}
